Generate keycodes to header
